\name{abbreviate.run.id.ED}
\alias{abbreviate.run.id.ED}
\title{Abbreviate run id to ed limits}
\usage{
  abbreviate.run.id.ED(run.id)
}
\arguments{
  \item{run.id}{string indicating nature of the run}
}
\description{
  As is the case with ED, input files must be <32
  characters long. this function abbreviates run.ids for
  use in input files. Depreciated.
}
\author{
  unknown
}

