% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{benchdat}
\alias{benchdat}
\title{"benchden" Dataset}
\format{\itemize{
   \item x random deviates
   \item density label for the density the deviates belong to
   \item dennum number for the density the deviates belong to (corresponds to the \code{dnum} parameter in benchden's \code{\link[benchden]{rberdev}})
}}
\usage{
benchdat
}
\description{
5000 random deviates from each of the 28 benchmark densities available in the "benchden" package
}
\keyword{data}

