% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{benchden}
\alias{benchden}
\title{"benchden" Dataset}
\format{\itemize{
   \item x random deviates
   \item density label for the density the deviates belong to
   \item dennum number for the density the deviates belong to (corresponds to the \code{dnum} parameter in benchden's \code{rberdev})
}}
\usage{
benchden
}
\description{
5000 random deviates from each of the 28 benchmark densities available in the "benchden" package
}
\keyword{data}

