% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ed2in.R
\name{between}
\alias{between}
\title{Check if value is between (inclusive) a range}
\usage{
between(x, lower, upper)
}
\arguments{
\item{x}{Value to check}

\item{lower}{Lower limit}

\item{upper}{Upper limit}
}
\description{
Check if value is between (inclusive) a range
}
