\name{calc.ed.specific}
\alias{calc.ed.specific}
\title{Calculate ED specific variables from other database traits}
\usage{
  calc.ed.specific(trait.samples)
}
\arguments{
  \item{trait.samples}{a matrix or dataframe of samples
  from the trait distribution}
}
\value{
  matrix or dataframe with derived ED specific traits
}
\description{
  Calculate ED specific variables from other database
  traits
}
\author{
  Shawn Serbin, David LeBauer, Carl Davidson
}

