% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ed2in.R
\name{check_ed2in}
\alias{check_ed2in}
\title{Check ED2IN}
\usage{
check_ed2in(ed2in)
}
\arguments{
\item{ed2in}{Named list of ED2IN tag-value pairs. See \link{read_ed2in}.}
}
\description{
Check the basic structure of \code{ed2in} object, as well as consistency among
arguments (e.g. run dates and coordinates are within the range of vegetation
and meteorology data).
}
