% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ed_metheader.R
\name{check_ed_metfile}
\alias{check_ed_metfile}
\title{Check individual ED metfile}
\usage{
check_ed_metfile(metfile, variables)
}
\arguments{
\item{metfile}{Path to meteorology file}

\item{variables}{Variables table from [ed_metheader][read_ed_metheader] object}
}
\value{
`NULL`, invisibly, if successful or throw an error
}
\description{
Check individual ED metfile
}
