% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ed_metheader.R
\name{check_ed_metheader}
\alias{check_ed_metheader}
\title{Check ED met header object}
\usage{
check_ed_metheader(ed_metheader, check_files = TRUE)
}
\arguments{
\item{ed_metheader}{ED meteorology header object (see [read_ed_metheader])}

\item{check_files}{Logical. If `TRUE`, perform basic diagnostics on met 
files as well.}
}
\description{
Check that the object has all components, and throw an error if anything is 
wrong. Optionally, do some basic checks of actualy meteorology files as 
well.
}
