% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_veg.R
\name{create_css}
\alias{create_css}
\alias{create_pss}
\alias{create_site}
\title{Create css, pss, and site files from examples}
\usage{
create_css(input, check = TRUE)

create_pss(input, check = TRUE)

create_site(input, check = TRUE)
}
\arguments{
\item{input}{Named \code{list} or \code{data.frame} containing columns to replace in examples}

\item{check}{Logical. If \code{TRUE} (default), also check files for validity.}
}
\value{
css, pss, or site object (\code{data.frame}, possibly with attributes)
}
\description{
Create css, pss, and site files from examples
}
