% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_veg.R
\name{create_ed_veg}
\alias{create_ed_veg}
\title{Create full ED vegetation input object}
\usage{
create_ed_veg(css, pss, site, latitude, longitude, check = TRUE, ...)
}
\arguments{
\item{css}{\link[=read_css]{css} object (\code{data.frame})}

\item{pss}{\link[=read_pss]{pss} object (\code{data.frame})}

\item{site}{\link[=read_site]{site} object (\code{data.frame})}

\item{latitude}{Latitude coordinate}

\item{longitude}{Longitude coordinate}

\item{check}{Logical. If \code{TRUE} (default), also check files for validity.}

\item{...}{Additional objects to store in list}
}
\description{
Create full ED vegetation input object
}
