% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_edi.R
\name{download_edi}
\alias{download_edi}
\title{Download ED inputs}
\usage{
download_edi(directory)
}
\arguments{
\item{directory}{Target directory for unzipping files. Will be created if it 
doesn't exist.}
}
\value{
`TRUE`, invisibly
}
\description{
Download and unzip common ED inputs from a public Open Science Framework 
(OSF) repository (https://osf.io/b6umf). Inputs include the Olson Global 
Ecosystems (OGE) database (`oge2OLD`) and the `chd` and `dgd` databases.
}
\details{
The total download size around 28 MB.
}
