% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed.R
\name{ed}
\alias{ed}
\title{Fit ed model to data}
\usage{
ed(x, k = 10, disjoint = TRUE, degree = 2, span = 0.3, xgrid = 500,
  bounds = c(min(x), max(x)), f = NULL, family = "gaussian",
  normalize = FALSE, delta = 0, lower = NULL, control = loess.control())
}
\arguments{
\item{x}{data, a vector}

\item{k}{the gap width (in number of observations) with which to compute the preliminary estimates}

\item{disjoint}{should non-overlapping gaps be computed? (default \code{TRUE})}

\item{degree}{degree of loess fitting  (see \code{\link{loess}})}

\item{span}{span of loess fitting (see \code{\link{loess}})}

\item{xgrid}{number of equally-spaced points along the support of \code{x} at which to compute the fit}

\item{bounds}{bounds at which to fit the density (see details)}

\item{f}{a function providing a true density or hypothesized density, with which the ed estimate can be compared (optional)}

\item{family}{loess parameter (\code{\link{loess}})}

\item{normalize}{should the resulting density be normalized so that it integrates to one?}

\item{delta}{grid augmentation parameter (experimental).  A value of 0 (default) disables grid augmentation.}

\item{lower}{grid augmentation parameter (experimental)}

\item{control}{loess parameter (see \code{\link{loess.control}})}
}
\value{
a list with a lot of things (to be documented...).  For now, look at str(result) to get an idea.
}
\description{
Calculate preliminary ed density estimates and fit ed model to data using loess
}
\details{
bounds...
}
\note{
This function is provided as a convenience, but often you may want to simply compute the preliminary estimates using \code{\link{ed_pre}} and iteratively figure out how to fit the preliminary estimates with whatever nonparametric method you like.
}
\author{
Ryan Hafen
}
\seealso{
\code{\link{ed_pre}}, \code{\link{ed_plot}}
}

