% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed_varlookup.R
\name{ed.var}
\alias{ed.var}
\title{Lookup function for translating commonly used ED variables
returns out list, readvar variables to read from file, expr if any derivation is needed}
\usage{
ed.var(varname)
}
\description{
Lookup function for translating commonly used ED variables
returns out list, readvar variables to read from file, expr if any derivation is needed
}
