% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ed2in_dates.R
\name{ed2in2time}
\alias{ed2in2time}
\title{Convert ED2IN \code{ITIMEA/Z} string to hour and minute}
\usage{
ed2in2time(itimea)
}
\arguments{
\item{itimea}{ED2IN time string, e.g. "1200"}
}
\value{
List containing numeric values of hour and minute
}
\description{
Convert ED2IN \code{ITIMEA/Z} string to hour and minute
}
