% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed2in_set_value.R
\name{ed2in_set_value}
\alias{ed2in_set_value}
\title{Substitute single value into ED2IN}
\usage{
ed2in_set_value(tag, value, ed2in, modstring = "Modified by PEcAn")
}
\arguments{
\item{tag}{ED2IN tag, e.g. "SFILIN"}

\item{value}{Value for setting the tag}

\item{ed2in}{Character vector containing ED2IN file, usually generated by
\code{readLines}.}
}
\description{
Given a tag, find that tag in the provided ED2IN character
vector and change its value. If the tag is not found, do nothing.
}
\author{
Alexey Shiklomanov
}
