% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed2in_set_value.R
\name{ed2in_set_value_list}
\alias{ed2in_set_value_list}
\title{Substitute list of tag-value pairs into ED2IN}
\usage{
ed2in_set_value_list(tag_val_list, ed2in, ...)
}
\arguments{
\item{tag_val_list}{Named list of tag-value pairs. List names should match
ed2in tags.}

\item{ed2in}{Character vector containing ED2IN file, usually generated by
\code{readLines}.}

\item{...}{Other parameters to \code{ed2in_set_value}}
}
\author{
Alexey Shiklomanov
}
