% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA.helpers.ED2.R
\name{ed2in_sub}
\alias{ed2in_sub}
\title{Substitute values into ED2IN}
\usage{
ed2in_sub(tag, value, ed2in)
}
\arguments{
\item{tag}{ED2IN tag, e.g. "NL%SFILIN"}

\item{value}{Value for setting the tag}

\item{ed2in}{Character vector containing ED2IN file, usually generated by 
\code{readLines}.}
}
\author{
Alexey Shiklomanov
}

