% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ed2in.R
\docType{data}
\name{ed2in_tag_rxp}
\alias{ed2in_tag_rxp}
\title{Regular expressions for ED2IN tags and values}
\format{An object of class \code{character} of length 1.}
\usage{
ed2in_tag_rxp
}
\description{
Regular expressions used to match ED2IN tags (first captured group) and 
values (second captured group).
}
\keyword{datasets}
