% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ed_plot}
\alias{ed_plot}
\title{Log density plot using ed preliminary estimates}
\usage{
ed_plot(x, data = NULL, panel = "panel.ed", prepanel = "prepanel.ed",
  groups = NULL, ylab = "Log Density", ...)
}
\arguments{
\item{x}{either a formula (e.g. "~ data | condvar") or a numeric vector for which to calculate and plot ed preliminary estimates}

\item{data}{if \code{x} is a formula, an optional data source (usually a data frame) in which variables are to be evaluated (see \code{\link{xyplot}} for details).}

\item{panel}{a function, called once for each panel, that uses the packet (subset of panel variables) corresponding to the panel to create a display. The default panel function is \code{\link{panel.ed}} and is documented separately, and has arguments that can be used to customize its output in various ways. Such arguments can usually be directly supplied to the high-level function.}

\item{prepanel}{See \code{\link{xyplot}}.}

\item{groups}{See \code{\link{xyplot}}.}

\item{ylab}{See \code{\link{xyplot}}.}

\item{\ldots}{further arguments. See corresponding entry in \code{\link{xyplot}} for non-trivial details.}
}
\value{
An object of class "trellis".
}
\description{
Plot the log density of data using ed preliminary estimates
}
\examples{
# plot benchden data
ed_plot(~ x | density, data = benchden,
   scales = list(relation = "free", draw = FALSE),
   as.table = TRUE
)

# generate some data
n <- 5000
x <- rnorm(3 * n, mean = rep(c(0, 2, 4), each = n), sd = rep(1:3, each = n))
dd <- data.frame(x = x, cond = rep(1:3, each = n))

# estimate/plot all the data
ed_plot(~ x, data = dd, xlab = "data", aspect = 1)

# estimate/plot by conditioning variable
ed_plot(~ x | cond, data = dd,
  xlab = "data", aspect = 1,
  as.table = TRUE, layout = c(3, 1))

# with lines
ed_plot(~ x | cond, data = dd, kk = 15,
  panel = function(x, ...) {
    panel.grid(h = -1, v = -1)
    panel.ed(x, ...)
    ss <- seq(-8, 15, length = 300)
    panel.lines(ss, log(dnorm(ss, mean = mean(x), sd = sd(x))), col = "black")
  },
  type = c("p", "g"),
  aspect = 1,
  as.table = TRUE,
  layout = c(3, 1)
)

# using "groups"
ed_plot(~ x, data = dd, groups = cond, alpha = 0.75)
}
\author{
Ryan Hafen
}
\seealso{
\code{\link{densityplot}}
}

