% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed_raw.R
\name{ed_raw}
\alias{ed_raw}
\title{Calculate ed Raw Estimates}
\usage{
ed_raw(x, k = 10, disjoint = TRUE, f = NULL)
}
\arguments{
\item{x}{data, a vector}

\item{k}{the gap width (in number of observations) with which to compute the raw estimates}

\item{disjoint}{should non-overlapping gaps be computed? (default TRUE)}

\item{f}{a function providing a true density or hypothesized density, with which the ed estimate can be compared (optional)}
}
\value{
a data frame with: \code{x}, the location at which the raw estimate was calculated; \code{raw} the raw log density estimate
}
\description{
Calculate ed raw estimates
}
\note{
It is advised to stick with \code{disjoint=TRUE} since this will ensure that the errors of subsequent fitting of the raw estimates will be independent.
}
\examples{
x <- rnorm(1000)
x_raw <- ed_raw(x)
plot(x_raw)
# overlay actual log density
ss <- seq(-4, 4, length = 200)
lines(ss, log(dnorm(ss)))
}
\author{
Ryan Hafen
}
\seealso{
\code{\link{ed}}, \code{\link{ed_plot}}
}

