% Generated by roxygen2 (4.0.2): do not edit by hand
\name{panel.edPlot}
\alias{panel.edPlot}
\alias{prepanel.edPlot}
\title{Panel / Prepanel Functions for edPlot}
\usage{
panel.edPlot(x, kk = 10, lims = TRUE, lwd = 2, type = "p",
  groups = NULL, ...)

prepanel.edPlot(x, kk = 10, groups = NULL, ...)
}
\arguments{
\item{x}{data, a vector}

\item{kk}{the gap width (in number of observations) with which to compute the preliminary estimates}

\item{lims,lwd,type,groups,\ldots}{parameters passed to \code{panel.xyplot}}
}
\description{
Panel / Prepanel Functions for edPlot
}

