% Generated by roxygen2 (4.0.2): do not edit by hand
\name{edPre}
\alias{edPre}
\title{Calculate ed Preliminary Estimates}
\usage{
edPre(x, k = 10, disjoint = TRUE, f = NULL)
}
\arguments{
\item{x}{data, a vector}

\item{k}{the gap width (in number of observations) with which to compute the preliminary estimates}

\item{disjoint}{should non-overlapping gaps be computed? (default TRUE)}

\item{f}{a function providing a true density or hypothesized density, with which the ed estimate can be compared (optional)}
}
\value{
a data frame with: \code{x}, the location at which the preliminary estimate was calculated; \code{pre} the preliminary log density estimate
}
\description{
Calculate ed preliminary estimates
}
\note{
It is advised to stick with \code{disjoint=TRUE} since this will ensure that the errors of subsequent fitting of the preliminary estimates will be independent.
}
\examples{
x <- rnorm(1000)
xPre <- edPre(x)
plot(xPre)
# overlay actual log density
ss <- seq(-4, 4, length = 200)
lines(ss, log(dnorm(ss)))
}
\author{
Ryan Hafen
}
\seealso{
\code{\link{ed}}, \code{\link{edPlot}}
}

