% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_veg.R
\docType{data}
\name{example_css}
\alias{example_css}
\alias{example_pss}
\alias{example_site}
\title{Example css, pss, and site objects}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 10 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 14 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 7 columns.
}
\usage{
example_css

example_pss

example_site
}
\description{
Example css, pss, and site objects
}
\keyword{datasets}
