% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.ED2.R
\name{extract_pfts}
\alias{extract_pfts}
\title{Extract pft numbers from settings$pfts}
\usage{
extract_pfts(pfts)
}
\arguments{
\item{pfts}{settings$pfts}
}
\value{
named numeric vector
}
\description{
A helper function to extract a named vector of pft numbers from
\code{settings$pfts}.  Will use pft numbers in \code{settings} if they exist, otherwise
it'll match using the \code{pftmapping} dataset
}
