% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.model.output.ED2}
\alias{get.model.output.ED2}
\title{Retrieve ED2 HDF model output from local or remote server}
\usage{
get.model.output.ED2(settings)
}
\description{
Function to retrieve ED2 HDF model output from local or remote server
}
\author{
Shawn Serbin

David LeBauer
}

