% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA.helpers.ED2.R
\name{get_configxml.ED2}
\alias{get_configxml.ED2}
\title{Get ED2 config.xml file}
\usage{
get_configxml.ED2(rundir, runid)
}
\arguments{
\item{rundir}{Model run directory. Usually \code{<workflowID>/run}}

\item{runid}{PEcAn run ID}
}
\description{
Get ED2 config.xml file
}
\author{
Alexey Shiklomanov
}
