% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ed2in_dates.R
\name{get_ed2in_dates}
\alias{get_ed2in_dates}
\title{Extract sequence of dates from ED2IN file}
\usage{
get_ed2in_dates(ed2in)
}
\arguments{
\item{ed2in}{Named list of ED2IN tag-value pairs. See \link{read_ed2in}.}
}
\value{
Vector of dates from start date to end date by 1 day
}
\description{
Extract sequence of dates from ED2IN file
}
