% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_met_dates.R
\name{get_met_dates}
\alias{get_met_dates}
\title{Get meteorology dates}
\usage{
get_met_dates(ed_metheader)
}
\arguments{
\item{ed_metheader}{ED meteorlogy header object (see \link{read_ed_metheader})}
}
\value{
Vector of dates for a run
}
\description{
Figure out the dates for which a given meteorology is available by parsing
the matching file names.
}
