% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA.helpers.ED2.R
\name{get_restartfile.ED2}
\alias{get_restartfile.ED2}
\title{Get ED history restart file path}
\usage{
get_restartfile.ED2(mod_outdir, runid, file.time)
}
\arguments{
\item{mod_outdir}{Directory where PEcAn stores ensemble outputs. Usually
\code{<workflowID>/out}}

\item{runid}{PEcAn run ID}

\item{file.time}{Start or end time from SDA analysis}
}
\description{
Get ED history restart file path
}
\author{
Alexey Shiklomanov
}
