% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.helpers.ED2.R
\name{list.files.nodir}
\alias{list.files.nodir}
\title{List only files in a directory}
\usage{
list.files.nodir(path, ...)
}
\arguments{
\item{path}{a character vector of full path names; the default
    corresponds to the working directory, \code{\link{getwd}()}.  Tilde
    expansion (see \code{\link{path.expand}}) is performed.  Missing
    values will be ignored.}
}
\description{
List only files in a directory
}
\author{
Alexey Shiklomanov
}
