% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met2model.ED2.R
\name{met2model.ED2}
\alias{met2model.ED2}
\title{met2model wrapper for ED2}
\usage{
met2model.ED2(in.path, in.prefix, outfolder, start_date, end_date,
  lst = 0, lat = NA, lon = NA, overwrite = FALSE,
  verbose = FALSE, leap_year = TRUE, ...)
}
\arguments{
\item{in.path}{location on disk where inputs are stored}

\item{in.prefix}{prefix of input and output files}

\item{outfolder}{location on disk where outputs will be stored}

\item{start_date}{the start date of the data to be downloaded (will only use the year part of the date)}

\item{end_date}{the end date of the data to be downloaded (will only use the year part of the date)}

\item{lst}{timezone offset to GMT in hours}

\item{overwrite}{should existing files be overwritten}

\item{verbose}{should the function be very verbose}

\item{leap_year}{Enforce Leap-years? If set to TRUE, will require leap years to have 366 days. If set to false, will require all years to have 365 days. Default = TRUE.}
}
\description{
If files already exist in 'Outfolder', the default function is NOT to
overwrite them and only gives user the notice that file already exists. If
user wants to overwrite the existing files, just change overwrite statement
below to TRUE.
}
