% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ed_metheader.R
\docType{data}
\name{met_flag_description}
\alias{met_flag_description}
\title{Description of meteorology flags}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 2 columns.}
\usage{
met_flag_description
}
\description{
Descriptions of ED met header variable flags.
}
\details{
`data.frame` with the following columns:
 - `flag` -- Numeric flag (in header file)
 - `flag_description` -- Description of flag
}
\keyword{datasets}
