% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ed_metheader.R
\docType{data}
\name{met_variable_description}
\alias{met_variable_description}
\title{Description of meteorology variables}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15 rows and 3 columns.
}
\usage{
met_variable_description
}
\description{
Helpful information about ED_MET_DRIVER files.
}
\details{
\code{data.frame} with the following columns:
\itemize{
\item \code{variable} -- Variable name
\item \code{description} -- Variable description
\item \code{unit} -- Variable unit (character, parse-able by \code{udunints2})
}
}
\keyword{datasets}
