\name{model2netcdf.ED2}
\alias{model2netcdf.ED2}
\title{Code to convert ED2's -T- HDF5 output into netCDF format}
\usage{
  model2netcdf.ED2(outdir, run.id)
}
\arguments{
  \item{outdir}{Location of SIPNET model output}

  \item{run.id}{Name of SIPNET model output file.}
}
\description{
  Modified from Code to convert ED2.1's HDF5 output into
  the NACP Intercomparison format (ALMA using netCDF)
}
\author{
  Michael Dietze
}

