% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_veg.R
\name{modify_df}
\alias{modify_df}
\title{Modify a reference \code{data.frame}}
\usage{
modify_df(input, base)
}
\arguments{
\item{input}{Named \code{list} or \code{data.frame} containing columns to replace in \code{base}}

\item{base}{Reference object to modify}
}
\value{
Modified \code{data.frame}
}
\description{
Wrapper around \code{modifyList} to allow expanding a \code{data.frame} by modifying
only a single column.
}
