% Generated by roxygen2 (4.0.1): do not edit by hand
\name{parse.history}
\alias{parse.history}
\title{Create a CSV from history.xml outputed by ED}
\usage{
parse.history(historyfile, outfile = "")
}
\arguments{
\item{historyfile}{filename of history file generated by ED.}

\item{outfile}{location where to write output, if no specified it will write to the console.}
}
\description{
This will generate the CSV file needed by write configs to write the
config.xml. This is a hack right now, all this information should be
in the PEcAn DB.
}
\author{
Rob Kooper
}

