% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA.helpers.ED2.R
\name{patch_cohort_index}
\alias{patch_cohort_index}
\title{Generate ED2 cohort to patch mapping vector}
\usage{
patch_cohort_index(nc)
}
\arguments{
\item{nc}{\code{ncdf4} object for ED history restart file.}
}
\description{
Generate a vector of integer indices for mapping ED state
cohort vectors onto patches, for instance for use with \code{tapply}.
}
\author{
Alexey Shiklomanov
}
