% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pftmapping}
\alias{pftmapping}
\title{Mapping of PEcAn PFT names to ED2 PFT numbers}
\format{
A data frame with 73 rows and 2 variables:
\describe{
\item{PEcAn}{PEcAn PFT names}
\item{ED}{ED2 PFT numbers}
...
}
}
\source{
\url{https://github.com/EDmodel/ED2/wiki/Plant-functional-types}
}
\usage{
pftmapping
}
\description{
A dataset matching PEcAn PFT names to ED PFT numbers.
}
\keyword{datasets}
