% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ed_veg.R
\name{prepare_ed_veg_filename}
\alias{prepare_ed_veg_filename}
\title{Format file name for ED vegetation inputs}
\usage{
prepare_ed_veg_filename(path_prefix, suffix, latitude = NULL,
  longitude = NULL)
}
\arguments{
\item{path_prefix}{Desired path and prefix (without latitude and longitude)}

\item{suffix}{Character string of filename suffix.}

\item{latitude}{Site latitude coordinate (default = \code{NULL})}

\item{longitude}{Site longitude coordinate (default = \code{NULL})}
}
\value{
Character string of full formatted file path
}
\description{
Adds the latitude and longitude, or checks if they are formatted correctly.
Then, splits the prefix into the directory and base name, appends the suffix
to the base name (adding a starting dot, if necessary), and returns the
filename as a character.
}
