% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ed2in.R
\name{print.ed2in}
\alias{print.ed2in}
\title{Print method for \code{ed2in}}
\usage{
print.ed2in(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Sets attributes to \code{NULL} before printing, so the output isn't as messy.
}
