% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{rmix}
\alias{dmix}
\alias{rmix}
\title{Mixture of Normal Distributions}
\usage{
rmix(n, mu = 0, sigma = 1, delta = 1)

dmix(x, mu, sigma, delta)
}
\arguments{
\item{n}{number of observations}

\item{mu}{vector of means}

\item{sigma}{vector of standard deviations}

\item{delta}{vector}

\item{x}{vector of quantiles}
}
\value{
\code{dmix} gives the density and \code{rnorm} generates random deviates
}
\description{
Density and random generation of a mixture of normal distributions
}
\details{
If a scalar is specified for any of the parameters, it will be replicated to meet the length of the other parameters.  If the \code{delta} values do not sum to 1, they will be normalized with a warning.
}
\examples{
mu <- c(0, 2, 3.5, 5, 8)
sigma <- c(0.5, 0.7, 0.3, 0.6, 1.4)
delta <- 0.2

x <- rmix(2000, mu, sigma, delta)
hist(x, breaks = 100, freq = FALSE)

ss <- seq(-2, 12, length = 200)
lines(ss, dmix(ss, mu, sigma, delta), col = "red")
}
\author{
Ryan Hafen
}

