% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.restart.ED2.R
\name{read.restart.ED2}
\alias{read.restart.ED2}
\title{State data assimilation read-restart for ED2}
\usage{
read.restart.ED2(outdir, runid, stop.time, settings, var.names, params)
}
\arguments{
\item{outdir}{output directory}

\item{runid}{run ID}

\item{stop.time}{year that is being read}

\item{settings}{PEcAn settings object}

\item{var.names}{var.names to be extracted}
}
\examples{
\dontrun{
  outdir <- "~/sda-hackathon/outputs"
  runid <- "99000000020"
  settings_file <- "outputs/pecan.CONFIGS.xml"
  settings <- PEcAn.settings::read.settings(settings_file)
  forecast <- read.restart.ED2(...)
}

}
\author{
Alexey Shiklomanov
}

