% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_veg.R
\name{read_css}
\alias{read_css}
\alias{read_pss}
\alias{read_site}
\title{Read individual css, pss, and site files}
\usage{
read_css(filepath, check = TRUE, ...)

read_pss(filepath, check = TRUE)

read_site(filepath, check = TRUE, ...)
}
\arguments{
\item{filepath}{Full path to css, pss, or site file}

\item{check}{Logical. If `TRUE` (default), [check][check_css] that file is valid.}

\item{...}{Additional arguments to [check functions][check_css].}
}
\value{
`data.frame` containing
}
\description{
Read files into objects usable by other PEcAn.ED2 utilities, and optionally check for errors.
}
