% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.ED2.R
\name{read_E_files}
\alias{read_E_files}
\title{Function for reading -E- files}
\usage{
read_E_files(
  yr,
  yfiles,
  efiles,
  outdir,
  start_date,
  end_date,
  pfts,
  settings = NULL,
  ...
)
}
\arguments{
\item{yr}{length 1 numeric vector; the year being processed}

\item{yfiles}{numeric vector of the years on the filenames, will be used to
matched efiles for that year}

\item{efiles}{character vector of names of E h5 files (e.g.
"analysis-E-1999-06-00-000000-g01.h5")}

\item{outdir}{directory where output will be written to}

\item{start_date}{Start time of the simulation}

\item{end_date}{End time of the simulation}

\item{pfts}{the \code{pfts} element from a pecan settings object}

\item{settings}{pecan settings object}

\item{...}{currently unused}
}
\value{
a list
}
\description{
This function reads in monthly output (-E- .h5 files) from ED2, does unit
conversions, and returns a list to be passed to \code{\link[=put_E_values]{put_E_values()}}.  Cohort
level variables (i.e. those ending in "_CO") are often (always?) in
per-plant units rather than per area. This function converts them to per
area using the plant density and patch area before converting units to PEcAn
standards.
}
\details{
if \code{settings} is provided, then values for missing arguments
for \code{outdir}, \code{start_date}, \code{end_date}, and \code{pfts} will be taken from it
}
