% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ed2in.R
\name{read_ed2in}
\alias{read_ed2in}
\title{Read ED2IN file to named list}
\usage{
read_ed2in(filename)
}
\arguments{
\item{filename}{Full path to ED2IN file}
}
\value{
Named list of \code{tag = value}
}
\description{
Parse an ED2IN file to a named list.
}
