% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed_veg.R
\name{read_ed_veg}
\alias{read_ed_veg}
\title{Read ED2 vegetation inputs}
\usage{
read_ed_veg(path_prefix, latitude = NULL, longitude = NULL, check = TRUE)
}
\arguments{
\item{path_prefix}{Full path and prefix to initial condition files.}

\item{latitude}{Run latitude (default = `NULL`). If `NULL`, deduced from file name.}

\item{longitude}{Run longitude (default = `NULL`). If `NULL`, deduced from file name.}

\item{check}{Whether or not to check css, pss, and site files for validity.
Default = `TRUE`.}
}
\value{
List containing `css`, `pss`, and `site` objects, `latitude` and 
`longitude`, and `orig_paths`, a list of paths to the original `css`, `pss`, 
and `site` files.
}
\description{
Read ED2 css, pss, and site files into a single ED input object.
}
