% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_restart.ED2.R
\name{read_restart.ED2}
\alias{read_restart.ED2}
\title{State data assimilation read-restart for ED2}
\usage{
read_restart.ED2(outdir, runid, stop.time, settings, var.names, params)
}
\arguments{
\item{outdir}{Output directory}

\item{runid}{Run ID}

\item{stop.time}{Year that is being read}

\item{settings}{PEcAn settings object}

\item{var.names}{Variable names to be extracted}

\item{params}{Any parameters required for state calculations}
}
\description{
State data assimilation read-restart for ED2
}
\examples{
\dontrun{
  outdir <- "~/sda-hackathon/outputs"
  runid <- "99000000020"
  settings_file <- "outputs/pecan.CONFIGS.xml"
  settings <- PEcAn.settings::read.settings(settings_file)
  forecast <- read_restart.ED2(...)
}

}
\author{
Alexey Shiklomanov, Istem Fer
}
