% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.configs.ed.R
\name{remove.config.ED2}
\alias{remove.config.ED2}
\title{Clear out old config and ED model run files.}
\usage{
remove.config.ED2(main.outdir = settings$outdir, settings)
}
\arguments{
\item{main.outdir}{ignored}

\item{settings}{PEcAn settings object}
}
\value{
nothing, removes config files as side effect
}
\description{
Clear out old config and ED model run files.
}
\author{
Shawn Serbin, David LeBauer, Alexey Shikomanov
}
