\name{remove.config}
\alias{remove.config}
\title{Clear out old config and ED model run files.}
\usage{
  remove.config(main.outdir, settings)
}
\value{
  nothing, removes config files as side effect
}
\description{
  Clear out old config and ED model run files.
}
\author{
  Shawn Serbin, David LeBauer
}

