\name{run.write.configs.ed}
\alias{run.write.configs.ed}
\title{Generate input ED2.2 model ED2IN and settings.xml files for PEcAn workflow}
\usage{
  run.write.configs.ed()
}
\description{
  Generate input ED2.2 model ED2IN and settings.xml files
  for PEcAn workflow
}
\author{
  Shawn Serbin

  David LeBauer

  Carl Davidson
}

