% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ed_singularity.R
\name{run_ed_singularity}
\alias{run_ed_singularity}
\title{Run ED singularity container}
\usage{
run_ed_singularity(
  img_path,
  ed2in_path,
  app = "ED",
  singularity_args = NULL,
  ...
)
}
\arguments{
\item{img_path}{Path to Singularity container (usually a \code{.simg} file)}

\item{ed2in_path}{Path to ED2IN file.}

\item{app}{Singularity "app" to run. Either "ED" or "EDR".}

\item{singularity_args}{Additional arguments to be passed to \verb{singularity run} (before)}

\item{Additional}{arguments to \link[base:system2]{base::system2}}
}
\description{
Uses \link[base:system2]{base::system2} to run ED or EDR via a Singularity container.
}
\details{
On some systems, to run Singularity properly, you will need to bind
additional paths. To do this, pass the arguments as a character vector to
\code{singularity_args}. For instance:\preformatted{bindpaths <- c("/scratch", "/data")
run_ed_singularity(..., singularity_args = paste("--bind", bindpaths))
}

By default, \link[base:system2]{base::system2} prints the output to the console. To store
standard ED output in a variable as a character vector, set \code{stdout = TRUE}.
To redirect all output to the variable, including GCC exceptions, use
\code{stderr = TRUE} (this will automatically set \code{stdout = TRUE} as well).
Output can also be redirected to a file via \code{stderr = "/path/to/file.log"}.
}
