% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.IC.ED2.R
\name{sample.IC.ED2}
\alias{sample.IC.ED2}
\title{Samples intial conditions for ED2}
\usage{
sample.IC.ED2(samples, format_name = "usda", start_date, pfts, ne, latitude,
  longitude, source, metadata, outfolder, host_info, inputs_path)
}
\arguments{
\item{samples}{a dataframe with columns: tree id, spp info, Measurement_Year, DBH, plot id (if applicable), could come from load_data}

\item{format_name}{how the spp info provided: usda, fia, latin_name, custom}

\item{start_date}{settings$run$start.date}

\item{pfts}{settings$pfts}

\item{ne}{number of ensembles requested}

\item{latitude}{settings$run$site$lat}

\item{longitude}{settings$run$site$lon}

\item{source}{the source of the samples, will appear as prefix in the IC filenames e.g. "HF_lyford.PalEON"}

\item{metadata}{any known metadata: age, area (to calculate density) etc.}

\item{outfolder}{settings$rundir}

\item{host_info}{host info, e.g. settings$host}

\item{inputs_path}{dir path to inputs, e.g. dirname(settings$run$inputs$css$path)}
}
\value{
IC
}
\description{
NOTE: almost all of the fcn args can be extracted from pecan settings list, just wnated to be explicit
}
\author{
Istem Fer
}
