% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAS.ED2.R
\name{SAS.ED2.param.Args}
\alias{SAS.ED2.param.Args}
\title{sets parameters and defaults for the ED2 semi-analytical spin-up}
\usage{
SAS.ED2.param.Args(
  decomp_scheme = 2,
  kh_active_depth = -0.2,
  decay_rate_fsc = 11,
  decay_rate_stsc = 4.5,
  decay_rate_ssc = 0.2,
  Lc = 0.049787,
  c2n_slow = 10,
  c2n_structural = 150,
  r_stsc = 0.3,
  rh_decay_low = 0.24,
  rh_decay_high = 0.6,
  rh_low_temp = 18 + 273.15,
  rh_high_temp = 45 + 273.15,
  rh_decay_dry = 12,
  rh_decay_wet = 36,
  rh_dry_smoist = 0.48,
  rh_wet_smoist = 0.98,
  resp_opt_water = 0.8938,
  resp_water_below_opt = 5.0786,
  resp_water_above_opt = 4.5139,
  resp_temperature_increase = 0.0757,
  rh_lloyd_1 = 308.56,
  rh_lloyd_2 = 1/56.02,
  rh_lloyd_3 = 227.15,
  yrs.met = 30,
  sm_fire = 0,
  fire_intensity = 0,
  slxsand = 0.33,
  slxclay = 0.33
)
}
\arguments{
\item{decomp_scheme}{Decomposition scheme specified in ED2IN}

\item{kh_active_depth}{Depth threshold for averaging soil moisture and temperature}

\item{decay_rate_fsc}{Fast soil carbon decay rate}

\item{decay_rate_stsc}{Structural soil carbon decay rate}

\item{decay_rate_ssc}{Slow soil carbon decay rate}

\item{Lc}{Used to compute nitrogen immpobilzation factor; ED default is 0.049787 (soil_respiration.f90)}

\item{c2n_slow}{Carbon to Nitrogen ratio, slow pool; ED Default 10.0}

\item{c2n_structural}{Carbon to Nitrogen ratio, structural pool. ED default 150.0}

\item{r_stsc}{Decomp param}

\item{rh_decay_low}{Param used for ED-1/CENTURY decomp schemes; ED default = 0.24}

\item{rh_decay_high}{Param used for ED-1/CENTURY decomp schemes; ED default = 0.60}

\item{rh_low_temp}{Param used for ED-1/CENTURY decomp schemes; ED default = 291}

\item{rh_high_temp}{Param used for ED-1/CENTURY decomp schemes; ED default = 318.15}

\item{rh_decay_dry}{Param used for ED-1/CENTURY decomp schemes; ED default = 12.0}

\item{rh_decay_wet}{Param used for ED-1/CENTURY decomp schemes; ED default = 36.0}

\item{rh_dry_smoist}{Param used for ED-1/CENTURY decomp schemes; ED default = 0.48}

\item{rh_wet_smoist}{Param used for ED-1/CENTURY decomp schemes; ED default = 0.98}

\item{resp_opt_water}{Param used for decomp schemes 0 & 3, ED default = 0.8938}

\item{resp_water_below_opt}{Param used for decomp schemes 0 & 3, ED default = 5.0786}

\item{resp_water_above_opt}{Param used for decomp schemes 0 & 3, ED default = 4.5139}

\item{resp_temperature_increase}{Param used for decomp schemes 0 & 3, ED default = 0.0757}

\item{rh_lloyd_1}{Param used for decomp schemes 1 & 4 (Lloyd & Taylor 1994); ED default = 308.56}

\item{rh_lloyd_2}{Param used for decomp schemes 1 & 4 (Lloyd & Taylor 1994); ED default = 1/56.02}

\item{rh_lloyd_3}{Param used for decomp schemes 1 & 4 (Lloyd & Taylor 1994); ED default = 227.15}

\item{yrs.met}{Number of years cycled in model spinup part 1}

\item{sm_fire}{Value to be used for SM_FIRE if INCLUDE_FIRE=2; defaults to 0 (fire off)}

\item{fire_intensity}{Value to be used for FIRE_PARAMTER; defaults to 0 (fire off)}

\item{slxsand}{Soil percent sand; used to calculate expected fire return interval}

\item{slxclay}{Soil percent clay; used to calculate expected fire return interval}
}
\description{
sets parameters and defaults for the ED2 semi-analytical spin-up
}
