% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAS.ED2.R
\name{SAS.ED2}
\alias{SAS.ED2}
\title{Use semi-analytic solution to accelerate model spinup}
\usage{
SAS.ED2(
  dir.analy,
  dir.histo,
  outdir,
  lat,
  lon,
  blckyr,
  prefix,
  treefall,
  param.args = SAS.ED2.param.Args(),
  sufx = "g01.h5"
)
}
\arguments{
\item{dir.analy}{Location of ED2 analysis files; expects monthly and yearly output}

\item{dir.histo}{Location of ED2 history files (for vars not in analy); expects monthly}

\item{outdir}{Location to write SAS .css & .pss files}

\item{lat}{site latitude; used for file naming}

\item{lon}{site longitude; used for file naming}

\item{blckyr}{Number of years between patch ages (aka blocks)}

\item{prefix}{ED2 -E- output file prefix}

\item{treefall}{Value to be used for TREEFALL_DISTURBANCE_RATE in ED2IN for full runs (disturbance on)}

\item{param.args}{ED2 parameter arguments (mostly soil biogeochem)}

\item{sufx}{ED2 out file suffix; used in constructing file names(default "g01.h5)}
}
\description{
This functions approximates landscape equilibrium steady state for vegetation and
soil pools using the successional trajectory of a single patch modeled with disturbance
off and the prescribed disturbance rates for runs (Xia et al. 2012 GMD 5:1259-1271).
}
\author{
Christine Rollinson, modified from original by Jaclyn Hatala-Matthes (2/18/14)
2014 Feb: Original ED SAS solution Script at PalEON modeling HIPS sites (Matthes)
2015 Aug: Modifications for greater site flexibility & updated ED
2016 Jan: Adaptation for regional-scale runs (single-cells run independently, but executed in batches)
2018 Jul: Conversion to function, Christine Rollinson July 2018
}
