% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ed2in.R
\name{tags2char}
\alias{tags2char}
\title{Format ED2IN tag-value list}
\usage{
tags2char(ed2in)
}
\arguments{
\item{ed2in}{Named list of ED2IN tag-value pairs. See [read_ed2in].}
}
\description{
Converts an `ed2in`-like list to an ED2IN-formatted character vector.
}
