% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_veg.R
\name{test_conditions}
\alias{test_conditions}
\title{Test list of conditions}
\usage{
test_conditions(conditions)
}
\arguments{
\item{conditions}{Vector of [expressions][base::expression].}
}
\value{
`NULL` (invisibly)
}
\description{
Evaluate a list of logical conditions, and throw an informative error if any 
are `FALSE`.
}
