% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.configs.ed.R
\name{write.config.ED2}
\alias{write.config.ED2}
\title{Write ED configuration files}
\usage{
write.config.ED2(trait.values, settings, run.id,
  defaults = settings$constants, check = FALSE, ...)
}
\arguments{
\item{trait.values}{Named list of trait values, with names corresponding to PFT}

\item{settings}{list of settings from pecan settings file}

\item{run.id}{id of run}

\item{defaults}{list of defaults to process. Default=settings$constants}

\item{check}{Logical. If \code{TRUE}, check ED2IN validity before running and
throw an error if anything is wrong (default = \code{FALSE})}
}
\value{
configuration file and ED2IN namelist for given run
}
\description{
Writes an xml and ED2IN config files for use with the Ecological Demography
model. Requires a pft xml object, a list of trait values for a single model
run, and the name of the file to create
}
\author{
David LeBauer, Shawn Serbin, Carl Davidson, Alexey Shiklomanov, Istem Fer
}
