% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.configs.ed.R
\name{write.config.jobsh.ED2}
\alias{write.config.jobsh.ED2}
\title{Write ED2 config.xml file}
\usage{
write.config.jobsh.ED2(settings, run.id)
}
\arguments{
\item{settings}{PEcAn settings list. For this function, need the following: 
run$host$rundir, run$host$outdir, run$host$scratchdir, 
run$host$clearscratch, run$jobtemplate, model$job.sh, run$host$job.sh, 
run$site$lat, run$site$lon, run$inputs$met$path, run$start.date, 
run$end.date, model$binary}

\item{run.id}{PEcAn run ID}
}
\value{
Character vector containing job.sh file
}
\description{
Function for writing job.sh file for ED2 runs
}
\details{
Refactored by Alexey Shiklomanov to allow use in PEcAn RTM module.
}
\author{
David LeBauer, Shawn Serbin, Carl Davidson, Alexey Shiklomanov
}

