% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.configs.ed.R
\name{write.config.xml.ED2}
\alias{write.config.xml.ED2}
\title{Write ED2 config.xml file}
\usage{
write.config.xml.ED2(defaults, settings, trait.values)
}
\arguments{
\item{defaults}{List of defaults to process}

\item{settings}{PEcAn settings file. Settings required for this script are: model$revision, model$config.header}

\item{trait.values}{}
}
\value{
R XML object containing full ED2 XML file
}
\details{
Refactored by Alexey Shiklomanov to allow use in PEcAn RTM module.
}
\author{
David LeBauer, Shawn Serbin, Carl Davidson, Alexey Shiklomanov
}

