% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.restart.ED2.R
\name{write.restart.ED2}
\alias{write.restart.ED2}
\title{Write ED2 restart file from SDA results}
\usage{
write.restart.ED2(outdir, runid, start.time, stop.time, settings, new.state,
  new.params = NULL, inputs)
}
\arguments{
\item{outdir}{output directory}

\item{runid}{run ID}

\item{settings}{PEcAn settings object}

\item{new.state}{analysis state vector}
}
\author{
Alexey Shiklomanov
}

