% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ed_veg.R
\name{write_ed_veg}
\alias{write_ed_veg}
\title{Write ED inputs to directory}
\usage{
write_ed_veg(ed_veg, path_prefix)
}
\arguments{
\item{ed_veg}{ED vegetation inputs object (see \link{read_ed_veg}).}

\item{path_prefix}{Desired path and prefix (without latitude and longitude)}
}
\value{
Named list (\code{css}, \code{pss}, \code{site}) of full file paths, invisibly
}
\description{
Write a complete \link[=read_ed_veg]{ED inputs object} to disk. \code{css}, \code{pss}, and
\code{site} files are automatically named and correctly formatted.
}
