% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_restart.ED2.R
\name{write_restart.ED2}
\alias{write_restart.ED2}
\title{Write ED2 restart file from SDA results}
\usage{
write_restart.ED2(outdir, runid, start.time, stop.time, settings, new.state,
  RENAME = TRUE, new.params, inputs)
}
\value{
TRUE if successful
}
\author{
Alexey Shiklomanov, Istem Fer
}
