% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ed_metheader.R
\name{check_ed_metfile}
\alias{check_ed_metfile}
\title{Check individual ED metfile}
\usage{
check_ed_metfile(metfile, variables)
}
\arguments{
\item{metfile}{Path to meteorology file}

\item{variables}{Variables table from \link[=read_ed_metheader]{ed_metheader} object}
}
\value{
\code{NULL}, invisibly, if successful or throw an error
}
\description{
Check individual ED metfile
}
