% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_met_dates.R
\name{dates_in_month}
\alias{dates_in_month}
\title{Get all the dates in a month}
\usage{
dates_in_month(date)
}
\description{
For a given date, figure out its month and return all of the dates for that
month.
}
