% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{panel.ed}
\alias{panel.ed}
\alias{prepanel.ed}
\title{Panel / Prepanel Functions for ed_plot}
\usage{
panel.ed(x, kk = 10, lims = TRUE, lwd = 2, type = "p", groups = NULL,
  ...)

prepanel.ed(x, kk = 10, groups = NULL, ...)
}
\arguments{
\item{x}{data, a vector}

\item{kk}{the gap width (in number of observations) with which to compute the preliminary estimates}

\item{lims, lwd, type, groups, \ldots}{parameters passed to \code{panel.xyplot}}
}
\description{
Panel / Prepanel Functions for ed_plot
}

