% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.helpers.ED2.R
\name{list.files.nodir}
\alias{list.files.nodir}
\title{List only files in a directory}
\usage{
list.files.nodir(path, ...)
}
\arguments{
\item{path}{directory to list}

\item{...}{arguments passed on to base::list.files}
}
\description{
Mostly useful when \code{recursive} and \code{full.names} are both FALSE:
The current implementation sets \code{full.names} internally, and for recursive
listings \code{list.files(..., include.dirs = FALSE)} is equivalent and faster.
}
\author{
Alexey Shiklomanov
}
