% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/met2model.ED2.R
\name{met2model.ED2}
\alias{met2model.ED2}
\title{met2model for ED2}
\usage{
met2model.ED2(in.path, in.prefix, outfolder, start_date, end_date, lst = 0,
  lat, lon, ..., overwrite = FALSE, verbose = FALSE)
}
\arguments{
\item{in.path}{location on disk where inputs are stored}

\item{in.prefix}{prefix of input and output files}

\item{outfolder}{location on disk where outputs will be stored}

\item{start_date}{the start date of the data to be downloaded (will only use the year part of the date)}

\item{end_date}{the end date of the data to be downloaded (will only use the year part of the date)}

\item{lst}{timezone offset to GMT in hours}

\item{overwrite}{should existing files be overwritten}

\item{verbose}{should the function be very verbose}
}
\description{
met2model wrapper for ED2
}

