% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.ED2.R
\name{model2netcdf.ED2}
\alias{model2netcdf.ED2}
\title{Code to convert ED2's -T- HDF5 output into netCDF format}
\usage{
model2netcdf.ED2(outdir, sitelat, sitelon, start_date, end_date,
  pft_names = NULL, dbh_breaks = 0)
}
\arguments{
\item{outdir}{Location of ED model output}

\item{sitelat}{Latitude of the site}

\item{sitelon}{Longitude of the site}

\item{start_date}{Start time of the simulation}

\item{end_date}{End time of the simulation}

\item{pft_names}{Names of PFTs used in the run}

\item{dbh_breaks}{Number of DBH size class breaks, default is 0 which represents a single DBH bin from 0 - Infinity cm}
}
\description{
Modified from Code to convert ED2.1's HDF5 output into the NACP Intercomparison format (ALMA using netCDF)
}
\author{
Michael Dietze, Shawn Serbin, Rob Kooper, Toni Viskari, Istem Fer
}
