% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.ED2.R
\name{model2netcdf.ED2}
\alias{model2netcdf.ED2}
\title{Code to convert ED2's -T- HDF5 output into netCDF format}
\usage{
model2netcdf.ED2(
  outdir,
  sitelat,
  sitelon,
  start_date,
  end_date,
  pfts,
  settings = NULL,
  process_partial = FALSE
)
}
\arguments{
\item{outdir}{Location of ED model output (e.g. a path to a single ensemble output)}

\item{sitelat}{Latitude of the site}

\item{sitelon}{Longitude of the site}

\item{start_date}{Start time of the simulation}

\item{end_date}{End time of the simulation}

\item{pfts}{a named vector of PFT numbers where the names are PFT names}

\item{settings}{pecan settings object}

\item{process_partial}{should failed runs be processed? Defaults to \code{FALSE}.
\code{TRUE} will generate .nc files for runs that have generated some, but not
all, of the expected outputs}
}
\description{
Modified from code to convert ED2's HDF5 output into the NACP
Intercomparison format (ALMA using netCDF)
}
\details{
if \code{settings} is provided, then values for missing arguments
\code{sitelat}, \code{sitelon}, \code{start_date}, \code{end_date}, and \code{pfts} will be taken
from it
}
\author{
Michael Dietze, Shawn Serbin, Rob Kooper, Toni Viskari, Istem Fer
}
