% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.ED2.R
\name{model2netcdf.ED2}
\alias{model2netcdf.ED2}
\title{Code to convert ED2's -T- HDF5 output into netCDF format}
\usage{
model2netcdf.ED2(
  outdir,
  sitelat,
  sitelon,
  start_date,
  end_date,
  pfts,
  settings = NULL
)
}
\arguments{
\item{outdir}{Location of ED model output}

\item{sitelat}{Latitude of the site}

\item{sitelon}{Longitude of the site}

\item{start_date}{Start time of the simulation}

\item{end_date}{End time of the simulation}

\item{pfts}{the \code{pfts} element from a pecan settings object}

\item{settings}{pecan settings object}
}
\description{
Modified from code to convert ED2's HDF5 output into the NACP
Intercomparison format (ALMA using netCDF)
}
\details{
if \code{settings} is provided, then values for missing arguments
will be taken from it
}
\author{
Michael Dietze, Shawn Serbin, Rob Kooper, Toni Viskari, Istem Fer
}
