% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.ED2.R
\name{model2netcdf.ED2}
\alias{model2netcdf.ED2}
\title{Code to convert ED2's -T- HDF5 output into netCDF format}
\usage{
model2netcdf.ED2(
  outdir,
  sitelat,
  sitelon,
  start_date,
  end_date,
  pft_names = NULL
)
}
\arguments{
\item{outdir}{Location of ED model output}

\item{sitelat}{Latitude of the site}

\item{sitelon}{Longitude of the site}

\item{start_date}{Start time of the simulation}

\item{end_date}{End time of the simulation}

\item{pft_names}{Names of PFTs used in the run, vector}
}
\description{
Modified from Code to convert ED2.1's HDF5 output into the NACP Intercomparison format (ALMA using netCDF)
}
\author{
Michael Dietze, Shawn Serbin, Rob Kooper, Toni Viskari, Istem Fer
}
