\name{model2netcdf.ED2}
\alias{model2netcdf.ED2}
\title{Code to convert ED2's -T- HDF5 output into netCDF format}
\usage{
  model2netcdf.ED2(outdir)
}
\arguments{
  \item{outdir}{Location of ED model output}
}
\description{
  #---------------------------------------------------------------------Modified
  from Code to convert ED2.1's HDF5 output into the NACP
  Intercomparison format (ALMA using netCDF)
}
\author{
  Michael Dietze, Shawn Serbin, Rob Kooper
}

