% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_ed2in.R
\name{modify_ed2in}
\alias{modify_ed2in}
\title{Modify an ED2IN object}
\usage{
modify_ed2in(
  ed2in,
  ...,
  veg_prefix = NULL,
  latitude = NULL,
  longitude = NULL,
  met_driver = NULL,
  start_date = NULL,
  end_date = NULL,
  EDI_path = NULL,
  output_types = NULL,
  output_dir = NULL,
  run_dir = NULL,
  runtype = NULL,
  run_name = NULL,
  include_these_pft = NULL,
  pecan_defaults = FALSE,
  add_if_missing = FALSE,
  check_paths = TRUE,
  .dots = list()
)
}
\arguments{
\item{ed2in}{list to modify}

\item{...}{Namelist arguments (see Description and Details)}

\item{veg_prefix}{Vegetation file prefix (\code{SFILIN}). If \code{lat} and \code{lon} are part of the prefix,}

\item{latitude}{Run latitude coordinate. If \code{veg_prefix} is also provided,
pass to \link{read_ed_veg}, otherwise set in ED2IN directly. Should be omitted if
\code{lat} and \code{lon} are already part of \code{veg_prefix}.}

\item{longitude}{Run longitude coordinate. If \code{veg_prefix} is also provided,
pass to \link{read_ed_veg}, otherwise set in ED2IN directly. Should be omitted if
\code{lat} and \code{lon} are already part of \code{veg_prefix}.}

\item{met_driver}{Path and filename of met driver header
(\code{ED_MET_DRIVER_DB})}

\item{start_date}{Run start date (\code{IMONTHA}, \code{IDATEA}, \code{IYEARA}, \code{ITIMEA})}

\item{end_date}{Run end date (\code{IMONTHZ}, \code{IDATEZ}, \code{IYEARZ} \code{ITIMEZ})}

\item{EDI_path}{Path to \code{EDI} directory, which often has the \code{VEG_DATABASE}
and \code{THSUMS_DATABASE} files.}

\item{output_types}{Character vector of output types (see Details)}

\item{output_dir}{Output directory, for \code{FFILOUT} (analysis) and \code{SFILOUT}
(history) files}

\item{run_dir}{Directory in which to store run-related config files (e.g. \code{config.xml}).}

\item{runtype}{ED initialization mode; either "INITIAL" or "HISTORY"}

\item{run_name}{Give the run an informative name/description. Sets
the ED2IN \code{EXPNME} tag. (default is \code{NULL})}

\item{include_these_pft}{Numeric vector describing the PFTs to
include in ED. Note that this is in addition to any PFTs
specified by the \code{config.xml} -- regardless of what this is set
to, those PFTs will be included, so if you want to only use PFTs
defined in \code{config.xml}, set this to \code{numeric(0)}. The default
(\code{NULL}) means to use whatever is already in the current ED2IN
file, which is usually all (1-17) of ED's PFTs.}

\item{pecan_defaults}{Logical. If \code{TRUE}, set common \code{ED2IN} defaults.}

\item{add_if_missing}{Logical. If \code{TRUE}, all-caps arguments not found in
existing \code{ed2in} list will be added to the end.  Default = \code{FALSE}.}

\item{check_paths}{Logical. If \code{TRUE} (default), for any parameters that
expect files, check that files exist and throw an error if they don't.}

\item{.dots}{A list of \code{...} arguments.}
}
\value{
Modified \code{ed2in} list object. See \link{read_ed2in}.
}
\description{
This is a convenience function for modifying an \code{ed2in} list object.
Arguments passed in all caps are assumed to be ED2IN namelist parameters and
are inserted directly into the \code{ed2in} list objects. Lowercase arguments are
defined explicitly (see "Parameters"), and those that do not match explicit
arguments will be ignored with a warning. Because the lowercase arguments
come with additional validity checks, they are recommended over modifying
the ED2IN file directly via uppercase arguments. For all lowercase
arguments, the default (\code{NULL}) means to use whatever is currently
in the input \code{ed2in}.
}
\details{
Namelist arguments are applied last, and will silently overwrite any
arguments set by special case arguments.

Namelist arguments can be stored in a list and passed in via the \code{.dots}
argument (e.g. \code{.dots = list(SFILIN = "/path/prefix_", ...)}), or using the
\verb{rlang::!!!} splicing operator. If both are provided, they will be spliced
together, with the \code{...} taking precedence.

For \code{output_types}, select one or more of the following:
\itemize{
\item "fast" -- Fast analysis; mostly polygon-level averages (\code{IFOUTPUT})
\item "daily -- Daily means (one file per day) (\code{IDOUTPUT})
\item "monthly" -- Monthly means (one file per month) (\code{IMOUTPUT})
\item "monthly_diurnal" -- Monthly means of the diurnal cycle (one file per
month) (\code{IQOUTPUT})
\item "annual" -- Annual (one file per year) (\code{IYOUTPUT})
\item "instant" -- Instantaneous fluxes, mostly polygon-level variables, one
file per year (\code{ITOUTPUT})
\item "restart" -- Restart file for HISTORY runs. (\code{ISOUTPUT})
\item "all" -- All output types
}
}
