% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.ED2.R
\name{put_E_values}
\alias{put_E_values}
\title{Function for put -E- values to nc_var list}
\usage{
put_E_values(yr, nc_var, out, lat, lon, begins, ends, pft_names, ...)
}
\description{
Function for put -E- values to nc_var list
}
