% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.ED2.R
\name{put_T_values}
\alias{put_T_values}
\title{Function for put -T- values to nc_var list}
\usage{
put_T_values(yr, nc_var, out, lat, lon, begins, ends, ...)
}
\description{
Function for put -T- values to nc_var list
}
