% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.ED2.R
\name{read_E_files}
\alias{read_E_files}
\title{Function for reading -E- files}
\usage{
read_E_files(yr, yfiles, efiles, outdir, start_date, end_date, pft_names,
  ...)
}
\arguments{
\item{yr}{the year being processed}

\item{yfiles}{the years on the filenames, will be used to matched efiles for that year}
}
\description{
Function for reading -E- files
}
\details{
e.g.     yr = 1999
yfiles = 1999 1999 1999 1999 1999 1999 1999 2000 2000 2000 2000
efiles = "analysis-E-1999-06-00-000000-g01.h5" "analysis-E-1999-07-00-000000-g01.h5"
"analysis-E-1999-08-00-000000-g01.h5" "analysis-E-1999-09-00-000000-g01.h5"
"analysis-E-1999-10-00-000000-g01.h5" "analysis-E-1999-11-00-000000-g01.h5"
"analysis-E-1999-12-00-000000-g01.h5" "analysis-E-2000-01-00-000000-g01.h5"
"analysis-E-2000-02-00-000000-g01.h5" "analysis-E-2000-03-00-000000-g01.h5"
"analysis-E-2000-04-00-000000-g01.h5"

pft_names  : character vector with names of PFTs
pft_names <- c("temperate.Early_Hardwood", "temperate.Late_Hardwood")
}
