% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ed_veg.R
\name{read_ed_veg}
\alias{read_ed_veg}
\title{Read ED2 vegetation inputs}
\usage{
read_ed_veg(path_prefix, latitude = NULL, longitude = NULL,
  check = TRUE)
}
\arguments{
\item{path_prefix}{Full path and prefix to initial condition files.}

\item{latitude}{Run latitude (default = \code{NULL}). If \code{NULL}, deduced from file name.}

\item{longitude}{Run longitude (default = \code{NULL}). If \code{NULL}, deduced from file name.}

\item{check}{Whether or not to check css, pss, and site files for validity.
Default = \code{TRUE}.}
}
\value{
List containing \code{css}, \code{pss}, and \code{site} objects, \code{latitude} and
\code{longitude}, and \code{orig_paths}, a list of paths to the original \code{css}, \code{pss},
and \code{site} files.
}
\description{
Read ED2 css, pss, and site files into a single ED input object.
}
