% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.ED2.R
\name{read_S_files}
\alias{read_S_files}
\title{S-file contents are not written to standard netcdfs but are used by read_restart
from SDA's perspective it doesn't make sense to write and read to ncdfs because ED restarts from history files}
\usage{
read_S_files(sfile, outdir, pfts, pecan_names = NULL, settings = NULL, ...)
}
\arguments{
\item{sfile}{history file name e.g. "history-S-1961-01-01-000000-g01.h5"}

\item{outdir}{path to run outdir, where the -S- file is}

\item{pfts}{the \code{pfts} element from a pecan settings object}

\item{pecan_names}{string vector, pecan names of requested variables, e.g. c("AGB", "AbvGrndWood")}

\item{settings}{pecan settings object}

\item{...}{currently unused}
}
\description{
S-file contents are not written to standard netcdfs but are used by read_restart
from SDA's perspective it doesn't make sense to write and read to ncdfs because ED restarts from history files
}
