% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.ED2.R
\name{read_S_files}
\alias{read_S_files}
\title{S-file contents are not written to standard netcdfs but are used by read_restart
from SDA's perspective it doesn't make sense to write and read to ncdfs because ED restarts from history files}
\usage{
read_S_files(sfile, outdir, pft_names, pecan_names = NULL)
}
\description{
S-file contents are not written to standard netcdfs but are used by read_restart
from SDA's perspective it doesn't make sense to write and read to ncdfs because ED restarts from history files
}
