% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.ED2.R
\name{read_T_files}
\alias{read_T_files}
\title{Function for reading -T- files}
\usage{
read_T_files(yr, yfiles, tfiles, outdir, start_date, end_date, ...)
}
\arguments{
\item{yr}{the year being processed}

\item{yfiles}{the years on the filenames, will be used to matched tfiles for that year}

\item{tfiles}{names of T files to be read}

\item{outdir}{directory where output will be written to}

\item{start_date}{start date in YYYY-MM-DD format}

\item{end_date}{end date in YYYY-MM-DD format}

\item{...}{additional arguments}
}
\description{
Function for reading -T- files
}
\details{
e.g.    yr = 1999
yfiles = 1999 2000
tfiles = "analysis-T-1999-00-00-000000-g01.h5" "analysis-T-2000-00-00-000000-g01.h5"
}
