% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.ED2.R
\name{read_T_files}
\alias{read_T_files}
\title{Function for reading -T- files}
\usage{
read_T_files(yr, yfiles, tfiles, outdir, start_date, end_date, ...)
}
\description{
yr     : the year being processed
yfiles : the years on the filenames, will be used to matched tfiles for that year
}
\details{
e.g.    yr = 1999
     yfiles = 1999 2000
     tfiles = "analysis-T-1999-00-00-000000-g01.h5" "analysis-T-2000-00-00-000000-g01.h5"
}
