% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/write.configs.ed.R
\name{remove.config.ED2}
\alias{remove.config.ED2}
\title{Clear out old config and ED model run files.}
\usage{
remove.config.ED2(main.outdir = settings$outdir, settings)
}
\value{
nothing, removes config files as side effect
}
\description{
Clear out old config and ED model run files.
}
\author{
Shawn Serbin, David LeBauer
}

