% Generated by roxygen2 (4.0.2): do not edit by hand
\name{start.runs.ED2}
\alias{start.runs.ED2}
\title{Start ED2 model runs}
\usage{
\method{start}{runs.ED2}(runid)
}
\arguments{
\item{runid}{the id of the run (folder in runs) to execute}
}
\description{
Start ED2 model runs on local or remote server
}
\author{
David LeBauer, Shawn Serbin, Carl Davidson
}

