\name{start.runs.ED2}
\alias{start.runs.ED2}
\title{Start ED2 model runs}
\usage{
  start.runs.ED2(runid)
}
\arguments{
  \item{runid}{the id of the run (folder in runs) to
  execute}
}
\description{
  Start ED2 model runs on local or remote server
}
\author{
  David LeBauer, Shawn Serbin, Carl Davidson
}

