% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.helpers.ED2.R
\name{translate_vars_ed}
\alias{translate_vars_ed}
\title{Function translating pecan vars to ED vars
var.names <- c("DBH", "AGB", "AbvGrndWood")}
\usage{
translate_vars_ed(varnames)
}
\description{
Function translating pecan vars to ED vars
var.names <- c("DBH", "AGB", "AbvGrndWood")
}
