% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg2model.ED2.R
\name{veg2model.ED2}
\alias{veg2model.ED2}
\title{Writes ED specific IC files}
\usage{
veg2model.ED2(outfolder, veg_info, start_date, new_site, source)
}
\arguments{
\item{outfolder}{where to write files}
}
\value{
filenames
}
\description{
Writes ED specific IC files
}
\author{
Istem Fer
}
