% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.configs.ed.R
\name{write.config.xml.ED2}
\alias{write.config.xml.ED2}
\title{Write ED2 config.xml file}
\usage{
write.config.xml.ED2(settings, trait.values, defaults = settings$constants)
}
\arguments{
\item{settings}{PEcAn settings file. Settings required for this script are: model$revision, model$config.header, constants}

\item{trait.values}{List of trait values with which to replace defaults}

\item{defaults}{List of defaults to process. Default = settings$constants}
}
\value{
R XML object containing full ED2 XML file
}
\details{
Refactored by Alexey Shiklomanov to allow use in PEcAn RTM module.
}
\author{
David LeBauer, Shawn Serbin, Carl Davidson, Alexey Shiklomanov
}

