% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ed_veg.R
\name{write_css}
\alias{write_css}
\alias{write_pss}
\alias{write_site}
\title{Write individual ED inputs}
\usage{
write_css(css, path_prefix, latitude = NULL, longitude = NULL)

write_pss(pss, path_prefix, latitude = NULL, longitude = NULL)

write_site(site, path_prefix, latitude = NULL, longitude = NULL)
}
\arguments{
\item{css}{css object (see \link{read_css})}

\item{path_prefix}{Desired path and prefix (without latitude and longitude)}

\item{latitude}{Site latitude coordinate (default = \code{NULL})}

\item{longitude}{Site longitude coordinate (default = \code{NULL})}

\item{pss}{pss object (see \link{read_pss})}

\item{site}{site object (see \link{read_site})}
}
\value{
Full file path as character, invisibly
}
\description{
Functions for writing css, pss, and site files from their respective objects.
}
\details{
Latitude and longitude coordinates will be converted directly to character,
without any changes to their precision. If they are \code{NULL} (default), the
function assumes that \code{lat} and \code{lon} are already in the \code{path_prefix}, and
if they are absent, the function will throw an error.
}
