% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ed_veg.R
\name{write_ed_veg}
\alias{write_ed_veg}
\title{Write ED inputs to directory}
\usage{
write_ed_veg(ed_veg, path_prefix)
}
\arguments{
\item{ed_veg}{ED vegetation inputs object (see [read_ed_veg]).}

\item{path_prefix}{Desired path and prefix (without latitude and longitude)}
}
\value{
Named list (`css`, `pss`, `site`) of full file paths, invisibly
}
\description{
Write a complete [ED inputs object][read_ed_veg] to disk. `css`, `pss`, and 
`site` files are automatically named and correctly formatted.
}
