% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_restart.ED2.R
\name{write_restart.ED2}
\alias{write_restart.ED2}
\title{Write ED2 restart file from SDA results}
\usage{
write_restart.ED2(outdir, runid, start.time, stop.time, settings, new.state,
  new.params, inputs, RENAME = TRUE)
}
\arguments{
\item{start.time}{Time of current assimilation step}

\item{stop.time}{Time of next assimilation step}

\item{new.state}{Analysis state matrix returned by \code{sda.enkf}}
}
\value{
TRUE if successful
}
\author{
Alexey Shiklomanov, Istem Fer
}
